Update InquiryContainer dynamic import
