+++ shared/common/src/components/InquiryContainer/InquiryList/style.ts
@@ -1,0 +1,9 @@
+import styled from '@emotion/styled'
+
+export const ListWrapper = styled.div`
+  display: flex;
+  justify-content: center;
+`
+export const ListContainer = styled.div`
+  width: 100%;
+`
