+++ shared/common/src/components/LectureApplyList/index.tsx
